requirements.txt
# ════════════════════════════════════════════════════
#  KEN ClawdBot — Python Dependencies
# ════════════════════════════════════════════════════

# ── AI / LLM ──────────────────────────────────────
anthropic>=0.25.0
openai>=1.30.0
google-generativeai>=0.5.0

# ── Web Framework (API bridge) ────────────────────
flask>=3.0.0
flask-cors>=4.0.0

# ── Scheduling ────────────────────────────────────
apscheduler>=3.10.4
pytz>=2024.1

# ── Twitter / X ───────────────────────────────────
tweepy>=4.14.0

# ── Google APIs (YouTube, Drive, Sheets) ──────────
google-api-python-client>=2.130.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# ── Notion ────────────────────────────────────────
notion-client>=2.2.1

# ── Media / Content ───────────────────────────────
elevenlabs>=1.2.0
requests>=2.31.0
pillow>=10.3.0
moviepy>=1.0.3

# ── Web scraping & research ───────────────────────
beautifulsoup4>=4.12.0
lxml>=5.2.0
httpx>=0.27.0

# ── Database / Memory ─────────────────────────────
peewee>=3.17.0
# sqlite3 is stdlib

# ── Utils ─────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
rich>=13.7.0
loguru>=0.7.2
tenacity>=8.3.0
python-dateutil>=2.9.0
